Qt: Guard better on profile ZIP archives
The Qt documentation has to be read very carefully to see which
functions remove ".." and which don't. Add more explicit comments
noting that absolute paths are allowed on Windows because of an
error creating profile ZIP archives up until 3.2.5 (but maybe we
should stop as that's a no longer supported version.) After converting
absolute paths to relative, perform a directory traversal check on
all OSes.

Also, don't cast the returned uncompressed_size in the unz_file_info64
struct from a ZPOS64_T (uint64_t) to an int, to handle the case where
a ZIP64 archive has a file with a 64-bit file size that when cast to
a 32-bit signed integer becomes negative or otherwise less than 256 MiB.

Fixup 70c86bd39c74fc0052c9a212afc911b41d612388

Fix #21553

+++ ui/qt/models/profile_model.cpp
@@ -1071,10 +1071,13 @@ bool ProfileModel::exportProfiles(QString filename, QModelIndexList items, QStri
 }
 
 /* This check runs BEFORE the file has been unzipped! */
-bool ProfileModel::acceptFile(QString fileName, int fileSize)
+bool ProfileModel::acceptFile(QString fileName, uint64_t fileSize)
 {
     if (fileName.toLower().endsWith(".zip"))
         return false;
+
+    /* Should we skip entries that are just directories? I don't believe
+     * we add them to our archive and it may simplify the logic elsewhere. */
 
     /* Arbitrary maximum config file size accepted: 256MB */
     if (fileSize > 1024 * 1024 * 256)
